Cookie Smasher
This is a Cookie Clickers bot that I made. To use simply download PyAutoGui and run the code in a terminal. To use simply put Cookie Clickers in your browser and put the terminal in the middle of the screen. This is only a macro basically, but I do plan on using some image recognition in the future to improve this.
Links:
Cookie Clickers https://orteil.dashnet.org/cookieclicker/
PyAutoGui https://pypi.org/project/PyAutoGUI/